Initial Jupyter Notebook
Have consolidated the FIFA Data into one file and Real Data into one file. Have done some basic feature engineering. Have only selected the columns we need. For the FIFA dataset have added a column which contains the players' rating in the previous edition of FIFA and  a column which contains the season on which the game is based upon. 

For the Real Statistics dataset, have added a column which contains standard name and a column which contains are based upon and a column which contains the season on which the real stats are based upon.